Refactor

+++ src/helper/formatDate.js
@@ -1,6 +1,6 @@
 export const formatDate = (fullDate) => {
   const date = new Date(fullDate);
-  const options = { month: "long", day: "numeric", year: "numeric" };
+  const options = { month: "short", day: "numeric", year: "numeric" };
   return date.toLocaleDateString("en-US", options);
 };
